+++ astropy/timeseries/tests/test_common.py
@@ -1,0 +1,60 @@
+from numpy.testing import assert_equal
+
+from ... import units as u
+from ...table import Table, QTable, vstack
+from ...time import Time
+from ..sampled import SampledTimeSeries
+from ..binned import BinnedTimeSeries
+
+
+INPUT_TIME = Time(['2016-03-22T12:30:31', '2015-01-21T12:30:32', '2016-03-22T12:30:40'])
+PLAIN_TABLE = Table([[1, 2, 11], [3, 4, 1], [1, 1, 1]], names=['a', 'b', 'c'])
+
+
+
+class CommonTimeSeriesTests:
+
+    def test_stacking(self):
+        ts = vstack([self.series, self.series])
+        assert isinstance(ts, self.series.__class__)
+
+    def test_row_slicing(self):
+        ts = self.series[:2]
+        assert isinstance(ts, self.series.__class__)
+
+    def test_row_indexing(self):
+        self.series[0][self.time_attr] == Time('2015-01-21T12:30:32')
+        self.series[self.time_attr][0] == Time('2015-01-21T12:30:32')
+
+    def test_column_indexing(self):
+        assert_equal(self.series['a'], [1, 2, 11])
+
+    def test_column_slicing_notime(self):
+        tab = self.series['a', 'b']
+        assert not isinstance(tab, self.series.__class__)
+        assert isinstance(tab, QTable)
+
+    def test_add_column(self):
+        self.series['d'] = [1, 2, 3]
+
+
+class TestSampledTimeSeries(CommonTimeSeriesTests):
+
+    def setup_method(self, method):
+        self.series = SampledTimeSeries(time=INPUT_TIME, data=PLAIN_TABLE)
+        self.time_attr = 'time'
+
+    def test_column_slicing(self):
+        ts = self.series['time', 'a']
+        assert isinstance(ts, SampledTimeSeries)
+
+
+class TestBinnedTimeSeries(CommonTimeSeriesTests):
+
+    def setup_method(self, method):
+        self.series = BinnedTimeSeries(start_time=INPUT_TIME, bin_size=3 * u.s, data=PLAIN_TABLE)
+        self.time_attr = 'start_time'
+
+    def test_column_slicing(self):
+        ts = self.series['start_time', 'bin_size', 'a']
+        assert isinstance(ts, BinnedTimeSeries)
